+++ lib/commands/local/invoke.js
@@ -11,9 +11,8 @@ const { getDebugPort, getDebugIde } = require('../../debug');
 const getVisitor = require('../../visitor').getVisitor;
 
 const { detectTplPath, getTpl } = require('../../tpl');
-const { findServices, findFunctions } = require('../../definition');
 
-const { red } = require('colors');
+const { red, yellow } = require('colors');
 
 function parseInvokeName(invokeName) {
   let serviceName = null;
@@ -35,25 +34,20 @@ function parseInvokeName(invokeName) {
   return [serviceName, functionName];
 }
 
-function findFirstFunction(tpl) {
-  var firstInvokeName;
-  const resources = tpl.Resources;
-  for (let { serviceName, serviceRes } of findServices(resources)) {
-    for (let { functionName } of findFunctions(serviceRes)) {
-      firstInvokeName = serviceName + '/' + functionName;
-      break;
-    }
-  }
-  return firstInvokeName;
-}
-
 async function localInvoke(invokeName, tpl, debugPort, event, debugIde, tplPath) {
   debug(`invokeName: ${invokeName}`);
 
   if (!invokeName) {
-    invokeName = findFirstFunction(tpl);
+
+    invokeName = definition.findFirstFunction(tpl);
+
+    if(!invokeName) {
+      throw new Error(red(`Missing function definition in template.yml`)); 
+    }
+    
+    console.log(`\nMissing invokeName argument, Fun will use the first function ${yellow(invokeName)} as invokeName\n`);
   }
-  
+
   const [parsedServiceName, parsedFunctionName] = parseInvokeName(invokeName);
 
   debug(`parse service name ${parsedServiceName}, functionName ${parsedFunctionName}`);
@@ -154,4 +148,4 @@ async function invoke(invokeName, options) {
   }
 }
 
-module.exports = { invoke, findFirstFunction };
+module.exports = invoke;
