+++ lib/definition.js
@@ -184,9 +184,21 @@ function findHttpTriggersInTpl(tpl) {
   return httpTriggers;
 }
 
+function findFirstFunction(tpl) {
+  var firstInvokeName;
+  const resources = tpl.Resources;
+  for (let { serviceName, serviceRes } of findServices(resources)) {
+    for (let { functionName } of findFunctions(serviceRes)) {
+      firstInvokeName = serviceName + '/' + functionName;
+      break;
+    }
+  }
+  return firstInvokeName;
+}
+
 module.exports = {
   findFunctionInTpl, findHttpTriggersInTpl,
   findFunctionsInTpl, findNasConfigInService,
   findHttpTriggersInFunction, findServices,
-  findFunctions
+  findFunctions, findFirstFunction
 };
